+++ gcovr/writer/html.py
@@ -40,6 +40,8 @@ from ..coverage import (
     DecisionCoverageStat,
     DecisionCoverageSwitch,
     DecisionCoverageUncheckable,
+    LineCoverage,
+    CallCoverage,
     SummarizedStats,
 )
 
@@ -318,15 +320,15 @@ def print_html_report(covdata: CovData, output_file, options):
     high_threshold_line = options.html_high_threshold_line
     medium_threshold_branch = options.html_medium_threshold_branch
     high_threshold_branch = options.html_high_threshold_branch
+    exclude_calls = options.exclude_calls
     show_decision = options.show_decision
-    show_calls = options.show_calls
 
     data = {}
     root_info = RootInfo(options)
     data["info"] = root_info
 
     data["SHOW_DECISION"] = show_decision
-    data["SHOW_CALL"] = show_calls
+    data["EXCLUDE_CALLS"] = exclude_calls
     data["COVERAGE_MED"] = medium_threshold
     data["COVERAGE_HIGH"] = high_threshold
     data["LINE_COVERAGE_MED"] = medium_threshold_line
@@ -486,6 +488,7 @@ def print_html_report(covdata: CovData, output_file, options):
                 lines = formatter.highlighter_for_file(data["filename"])(
                     source_file.read()
                 )
+                ctr = 0
                 for ctr, line in enumerate(lines, 1):
                     data["source_lines"].append(
                         source_row(ctr, line, cdata.lines.get(ctr))
@@ -496,7 +499,9 @@ def print_html_report(covdata: CovData, output_file, options):
                     )
         except IOError as e:
             logger.warning(f'File {data["filename"]} not found: {repr(e)}')
-            for ctr in range(1, max_line_from_cdata):
+            # Python ranges are exclusive. We want to iterate over all lines, including
+            # that last line. Thus, we have to add a +1 to include that line.
+            for ctr in range(1, max_line_from_cdata + 1):
                 data["source_lines"].append(
                     source_row(
                         ctr,
@@ -544,7 +549,7 @@ def dict_from_stat(
     return data
 
 
-def source_row(lineno, source, line_cov):
+def source_row(lineno: int, source: str, line_cov: Optional[LineCoverage]) -> dict:
     linebranch = None
     linedecision = None
     linecall = None
@@ -601,7 +606,7 @@ def source_row_branch(branches):
     }
 
 
-def source_row_call(calls):
+def source_row_call(calls: Optional[CallCoverage]):
     if not calls:
         return None
 
@@ -628,7 +633,7 @@ def source_row_call(calls):
     }
 
 
-def source_row_decision(decision: DecisionCoverage) -> Optional[dict]:
+def source_row_decision(decision: Optional[DecisionCoverage]) -> Optional[dict]:
     if decision is None:
         return None