+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Breaking changes:
 
 - Remove not allowed attributes ``function-rate``, ``functions-covered`` and ``functions-valid``
   from cobertura report. (:issue:`671`)
+- Remove "noncode" entries in JSON reports. (:issue:`663`)
 
 New features and notable changes:
 
@@ -27,8 +28,13 @@ Bug fixes and small improvements:
 - Do not ignore returncode of `gcov`. (:issue:`653`)
 - Fix alphabetical sort of html report, for when there are symlinks. (:issue:`685`)
 - Handle :option:`--version` before parsing the configuration file. (:issue:`696`)
+- Fix reports of excluded coverage. (:issue:`409`, :issue:`503`, :issue:`663`)
+- Fix handling for nonexistent source code for HTML-details and Coveralls reports (:issue:`663`)
+- :ref:`Exclusion marker regions <exclusion markers>` now contain the STOP marker line (:issue:`663`)
 
 Documentation:
+
+- Add detailed reference for the JSON output format. (:issue:`663`)
 
 Internal changes: